requirements.txt
attrdict
flask
gevent
pyfiglet
pyyaml
termcolor